* sccp: (29 commits)
  Use existing bit
  Updated Windows build
  Fixed compilation error
  Remove debug code
  We need to check for the length here too, or we crash and no one likes that! :(
  * Implemented #65187 (exif_read_data/thumbnail: add support for stream resource) * ext/exif now uses FAST_ZPP
  Remove extraneous configure flag
  Revert "remove excessive checks and fix warnings"
  parametrize zip names
  Upgrade bundled PCRE to 8.41
  Updated NEWS file with LDAP changes
  Fixed removing all controls by passing an empty array to ldap_set_option
  Filled in NEWS file with ext/ldap last modifications
  change order, allow to build as shared extension
  restore file deleted by mistake in a merge commit
  Fix segfault in php_stream_context_get_option call
  remove excessive checks and fix warnings
  fix macro redifinitions
  fix symbol availability and ws
  Remove this for now, as not found
  ...

+++ ext/zlib/php_zlib.h
@@ -48,7 +48,7 @@ typedef struct _php_zlib_buffer {
 typedef struct _php_zlib_context {
 	z_stream Z;
 	char *inflateDict;
-	size_t status;
+	int status;
 	size_t inflateDictlen;
 	php_zlib_buffer buffer;
 } php_zlib_context;
